Seperate package from example
