+++ example/go.mod
@@ -1,0 +1,20 @@
+module example.com
+
+go 1.20
+
+require (
+	github.com/dyxj/go-tsp v0.0.0
+	gonum.org/v1/plot v0.0.0-20190410204940-3a5f52653745
+)
+
+require (
+	github.com/ajstarks/svgo v0.0.0-20181006003313-6ce6a3bcf6cd // indirect
+	github.com/fogleman/gg v1.3.0 // indirect
+	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
+	github.com/jung-kurt/gofpdf v1.0.3-0.20190309125859-24315acbbda5 // indirect
+	golang.org/x/exp v0.0.0-20190419195159-b8972e603456 // indirect
+	golang.org/x/image v0.0.0-20190417020941-4e30a6eb7d9a // indirect
+	gonum.org/v1/gonum v0.0.0-20190419091250-b869779d1d53 // indirect
+)
+
+replace github.com/dyxj/go-tsp v0.0.0 => ../
