+++ go-tsp.code-workspace
@@ -1,0 +1,11 @@
+{
+  "folders": [
+	{
+		"path": "."
+	},
+	{
+		"path": "example"
+	}
+],
+  "settings": {}
+}
